Updated service station selection
Updated service station selection such that a customer also checks whether each station is empty before determining which line to join based on wait line length.

This can be improved if we can figure out a way to correctly break the logic off into a function without causing problems for the environment.